src/canvas/templates/sl/{{cookiecutter.project_name}}/requirements.txt
antlr4-python3-runtime==4.9.3
filelock==3.18.0
fsspec==2025.5.1
hydra-core==1.3.2
Jinja2==3.1.6
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
packaging==25.0
pillow==11.2.1
PyYAML==6.0.2
sympy==1.14.0
torch==2.2.2
torchvision==0.17.2
tqdm==4.67.1
typing_extensions==4.14.0
antlr4-python3-runtime==4.9.3
filelock==3.18.0
fsspec==2025.5.1
hydra-core==1.3.2
jinja2==3.1.6
markupsafe==3.0.2
mpmath==1.3.0
networkx==3.5
numpy==1.26.4
omegaconf==2.3.0
packaging==25.0
pyyaml==6.0.2
sympy==1.14.0
torch==2.2.2
tqdm==4.67.1
typing-extensions==4.14.0
